# openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunRequest": {
        "properties": {
          "confirm": {
            "default": false,
            "description": "required: a program can command unbounded motion",
            "title": "Confirm",
            "type": "boolean"
          },
          "skip_validation": {
            "default": false,
            "description": "run without checking the program's motion targets. Needed for programs FWS cannot read statically -- point names, computed poses, arcs -- and recorded in the audit log with the reason the check was skipped",
            "title": "Skip Validation",
            "type": "boolean"
          },
          "validation_note": {
            "default": "",
            "description": "why validation was skipped; recorded in the audit log",
            "title": "Validation Note",
            "type": "string"
          }
        },
        "title": "RunRequest",
        "type": "object"
      },
      "SelectRequest": {
        "description": "Choose the program to run, and optionally start it in one call.",
        "properties": {
          "confirm": {
            "default": false,
            "description": "required when start=true: a program commands unbounded motion",
            "title": "Confirm",
            "type": "boolean"
          },
          "start": {
            "default": false,
            "description": "also run it immediately",
            "title": "Start",
            "type": "boolean"
          }
        },
        "title": "SelectRequest",
        "type": "object"
      },
...
      "fws__programs_api__UploadRequest": {
        "properties": {
          "content": {
            "description": "Lua source",
            "title": "Content",
            "type": "string"
          },
          "overwrite": {
            "default": false,
            "title": "Overwrite",
            "type": "boolean"
          }
        },
        "required": [
          "content"
        ],
        "title": "UploadRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "REST + WebSocket gateway for Fairino collaborative robots. NOTE: no endpoint here is an emergency stop. The physical E-stop is hardware only. /motion/stop is a functional stop of jog motion.",
    "title": "Fairino Web Services"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/execution/pause": {
      "post": {
        "operationId": "handler_api_v1_execution_pause_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-fws-control-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fws-Control-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Handler",
        "tags": [
          "programs"
        ]
      }
    },
    "/api/v1/execution/resume": {
      "post": {
        "operationId": "handler_api_v1_execution_resume_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-fws-control-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fws-Control-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Handler",
        "tags": [
          "programs"
        ]
      }
    },
    "/api/v1/execution/run": {
      "post": {
        "description": "Start the loaded program.\n\nA Lua program commands motion directly on the controller: the gateway's\njog bounds, step limits and IK pre-flight do not apply.",
        "operationId": "run_api_v1_execution_run_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-fws-control-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fws-Control-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run",
        "tags": [
          "programs"
        ]
      }
    },
...
    "/api/v1/poses/program": {
      "post": {
        "description": "Turn a list of named poses into a runnable Lua program.\n\nReturns the source. It is deliberately NOT uploaded or run here:\ngenerating a program is safe, and running one moves the arm, so they\nstay separate calls with separate gates. Upload it with\n`PUT /api/v1/programs/{name}`, check it with `.../validate`, then run\nit through `/api/v1/execution/run`.",
        "operationId": "generate_api_v1_poses_program_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgramBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Generate Api V1 Poses Program Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate",
        "tags": [
          "poses"
        ]
      }
    },
...
    "/api/v1/programs/{name}": {
      "delete": {
        "operationId": "delete_program_api_v1_programs__name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-fws-control-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fws-Control-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Program",
        "tags": [
          "programs"
        ]
      },
      "get": {
        "operationId": "download_program_api_v1_programs__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Program",
        "tags": [
          "programs"
        ]
      },
      "put": {
        "operationId": "upload_program_api_v1_programs__name__put",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-fws-control-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fws-Control-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/fws__programs_api__UploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Program",
        "tags": [
          "programs"
        ]
      }
    },
    "/api/v1/programs/{name}/load": {
      "post": {
        "operationId": "load_program_api_v1_programs__name__load_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-fws-control-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fws-Control-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Load Program",
        "tags": [
          "programs"
        ]
      }
    },
    "/api/v1/programs/{name}/select": {
      "post": {
        "description": "Load a program and optionally start it, with the same guards\nas run: motion lock, confirmation, non-faulted controller.",
        "operationId": "select_program_api_v1_programs__name__select_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-fws-control-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fws-Control-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Select Program",
        "tags": [
          "programs"
        ]
      }
    },
    "/api/v1/programs/{name}/validate": {
      "post": {
        "description": "Check a program's motion targets without running it.\n\nSolves each literal motion target backwards through the controller's\nkinematics against its soft limits. Read `unchecked` before trusting\n`safe_to_run`: point names and computed poses cannot be checked.",
        "operationId": "validate_program_api_v1_programs__name__validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Program",
        "tags": [
          "programs"
        ]
      }
    },
...
  }
}